In this version we have made some modification in order to optimize its performance. Additionally, we are working in a new interface to connect EDLUT with real robot. This version include new restriction in order to achieve real time simulation connected to real robots.

+++ branches/New_Integrator_with_CUDA/include/integration_method/RK45.h
@@ -103,7 +103,7 @@ class RK45 : public FixedStep {
 		 * \param NumberOfEcuation number of differential ecuation.
 		 * \param elapsed_time integration time step.
 		 */
-		virtual void NextDifferentialEcuationValue(int index, TimeDrivenNeuronModel * Model, float * NeuronState, double elapsed_time, int CPU_thread_index);
+		virtual void NextDifferentialEcuationValue(int index, TimeDrivenNeuronModel * Model, float * NeuronState, float elapsed_time, int CPU_thread_index);
 
 		/*!
 		 * \brief It prints the integration method info.
