+++ branches/New_Integrator_with_CUDA/include/integration_method/VariableStep.h
@@ -82,7 +82,7 @@ class VariableStep : public IntegrationMethod {
 		 * \param elapsed_time integration time step.
 		 * \param CPU_thread_index index of the OpenMP thread.
 		 */
-		virtual void NextDifferentialEcuationValue(int index, TimeDrivenNeuronModel * Model, float * NeuronState, double elapsed_time, int CPU_thread_index) = 0;
+		virtual void NextDifferentialEcuationValue(int index, TimeDrivenNeuronModel * Model, float * NeuronState, float elapsed_time, int CPU_thread_index) = 0;
 
 		/*!
 		 * \brief It prints the integration method info.
