+++ branches/New_Integrator_with_CUDA/src/integration_method/RK45.cpp
@@ -19,13 +19,13 @@
 
 
 RK45::RK45(int N_neuronStateVariables, int N_differentialNeuronState, int N_timeDependentNeuronState, int N_CPU_thread):FixedStep("RK45", N_neuronStateVariables, N_differentialNeuronState, N_timeDependentNeuronState, N_CPU_thread, false, false),
-	a1(25.0/216.0), a2(0), a3(1408.0/2565.0), a4(2197.0/4104.0), a5(-0.2),
-	b1(16.0/135.0), b2(0), b3(6656.0/12825.0), b4(28561.0/56430.0), b5(-0.18), b6(2.0/55.0),
-	c20(0.25), c21(0.25),
-	c30(0.375), c31(0.09375), c32(0.28125),
-	c40(12.0/13.0), c41(1932.0/2197.0), c42(-7200.0/2197.0), c43(7296.0/2197.0),
-	c51(439.0/216.0), c52(-8.0), c53(439.0/216.0), c54(-845.0/4104.0),
-	c60(0.5), c61(-8.0/27.0), c62(2), c63(-3544.0/2565.0), c64(1859.0/4104.0), c65(-0.275)
+	a1(25.0f/216.0f), a2(0), a3(1408.0f/2565.0f), a4(2197.0f/4104.0f), a5(-0.2f),
+	b1(16.0f/135.0f), b2(0), b3(6656.0f/12825.0f), b4(28561.0f/56430.0f), b5(-0.18f), b6(2.0f/55.0f),
+	c20(0.25f), c21(0.25f),
+	c30(0.375f), c31(0.09375f), c32(0.28125f),
+	c40(12.0f/13.0f), c41(1932.0f/2197.0f), c42(-7200.0f/2197.0f), c43(7296.0f/2197.0f),
+	c51(439.0f/216.0f), c52(-8.0f), c53(439.0f/216.0f), c54(-845.0f/4104.0f),
+	c60(0.5f), c61(-8.0f/27.0f), c62(2), c63(-3544.0f/2565.0f), c64(1859.0f/4104.0f), c65(-0.275f)
 {	
 	AuxNeuronState = new float [N_NeuronStateVariables*N_CPU_thread];
 	AuxNeuronState1 = new float [N_NeuronStateVariables*N_CPU_thread];
@@ -50,7 +50,7 @@ RK45::~RK45(){
 	free (epsilon);
 }
 		
-void RK45::NextDifferentialEcuationValue(int index, TimeDrivenNeuronModel * Model, float * NeuronState, double elapsed_time, int CPU_thread_index){
+void RK45::NextDifferentialEcuationValue(int index, TimeDrivenNeuronModel * Model, float * NeuronState, float elapsed_time, int CPU_thread_index){
 	float * offset_AuxNeuronState = AuxNeuronState+(N_NeuronStateVariables*CPU_thread_index);
 	float * offset_AuxNeuronState1 = AuxNeuronState1+(N_NeuronStateVariables*CPU_thread_index);
 	float * offset_AuxNeuronState2 = AuxNeuronState2+(N_NeuronStateVariables*CPU_thread_index);
